requirements.txt
transformers>=4.0.0
torch>=1.9.0
pypdf>=3.0.0
python-docx>=0.8.11
pandas>=1.3.0
lancedb>=0.1.0
numpy>=1.21.0
beautifulsoup4>=4.9.3
requests>=2.26.0
docling
python-dotenv>=0.19.0
tiktoken>=0.5.0